requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Pillow==8.3.1
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
WTForms==2.3.3
email-validator==1.1.3 